Updated nav-item active status

+++ storagevault/src/main/java/com/cs401/storagevault/model/repository/PricingRepository.java
@@ -7,8 +7,8 @@ import org.springframework.data.jpa.repository.Query;
 public interface PricingRepository extends JpaRepository<Pricing, Integer> {
 
     @Query(value = "SELECT monthlyPrice FROM pricing WHERE customerType = :customerType", nativeQuery = true)
-    int findByCustomerType(char customerType);
+    double findByCustomerType(char customerType);
 
     @Query(value = "SELECT yearlyPrice FROM pricing WHERE customerType = :customerType", nativeQuery = true)
-    int yearlySubscriptionPrice(char customerType);
+    double yearlySubscriptionPrice(char customerType);
 }
